move extractDataSourceConfiguration into a dedicated file
